defis/22_Proces_JK_Rowling/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
textblob>=0.17.1
nltk>=3.8.1
spacy>=3.7.2
spacy-langdetect>=0.1.2
unidecode>=1.3.7
python-slugify>=8.0.4
jupyter>=1.0.0
